[FIX] Click-and-hold should only pop up the context menu if performed on the action button.
This still presents the problem where, after clicking and holding, you must release the mouse button in order to select an item. This is due to the fact that the context menu is faked, not a system menu. The desired behavior may be doable with some serious work, but I'm going to leave it be for now.